requirements.txt
chromadb==0.4.13
elastic-transport==8.4.1
elasticsearch==8.10.1
faiss-cpu==1.7.4
fastapi==0.103.1
google-search-results==2.4.2
langchain==0.0.340
langchainhub==0.1.13
langsmith==0.0.66
llama-hub==0.0.41
llama-index==0.8.65
neo4j==5.13.0
numpy==1.26.0
openai==1.3.5
pydantic==1.10.13
pyngrok==7.0.1
pypdf==3.16.2
python-dotenv==0.19.2
streamlit==1.27.2
tiktoken==0.5.1
uvicorn==0.23.2
wikipedia==1.4.0
newspaper3k==0.2.8